requirements.txt
certifi==2020.12.5
chardet==3.0.4
click==8.0.1
dominate==2.6.0
Flask==2.0.1
Flask-Bootstrap4==4.0.2
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
idna==2.10
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
requests==2.25.1
SQLAlchemy==1.4.20
urllib3==1.26.3
visitor==0.1.3
Werkzeug==2.0.1
WTForms==2.2.1
gunicorn==20.1.0
